requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
Pillow
fastapi
uvicorn
timm
